Switch getBalance() method to using UUID (so it can actually get the balance)
To withdraw money, you need to know if you've got enough.
To know if you've got enough, you need to know your balance
To know your balance, you need to check in the NBT file
To check in the NBT file, you need to know the card owners UUID.

Players will not need a new debit card; the debit card will self-update
when the player holds it and add the CURRENTLY HOLDING PLAYER's UUID.
So, regardless of the original owner, the person holding the card will
now be the owner.
The name shown on the card will remain as the original. This will only
be a minor bug for existing and stolen debit cards; which I can't
imagine there are many of. Consider it a rare & unique item ;)