Fixing latency issues with balance updates
Balance is now refreshed after the GUI opening (on every button press,
to be precise), so there shouldn't be any issues with it updating.
Withdrawing money will now update your balance too, which is nice.

(This will not have a network impact. The GUI requests an update at the
same time as it tells the server to play the "Beep" sound, the only
information actually being sent is the balance update which shouldn't
be more than a few bytes for even your richest players.)