MAJOR BREAKING CHANGES TO ECON API
If you are using the econ api then I've just ruined your day.
All references to world are now removed, and called internally by the
NBT loader.

This on the plus side means you can now call any API methods with only a
player instance, no need for the world. However, it means anything using
the API will have to update to version 0.15 and above to continue to be
compatible.

The change was required by a mod utilizing the API which has not got a
world instance (As would be the case for generic things that maybe don't
exist within the world.)